crosstool-NG.sh.in: Don't make lots of symlinks to lib folder
For 4 different folders:

${CT_PREFIX_DIR}
${CT_SYSROOT_DIR}
${CT_SYSROOT_DIR}/usr
${CT_PREFIX_DIR}/${CT_TARGET}

.. symlinks from lib32 and lib64 to lib were created. This
is a lot of symlinks and is untidy.

They're likely not needed in any case, but for now, excluded
for the multilib case only.

The removal of these symlinks is avoided if they weren't
created too. In the non-multilib case, removal via rm -f
was changed to rm -rf so systems (Win32) that use cp -rf
instead of ln -s don't fail when rm -f gets called on a
directory. This change is minor so was slipped in here!

Signed-off-by: Ray Donnelly <mingw.android@gmail.com>